// src/lib/generated/repokit/types.ts
// Generated by repokit publish

export interface RepokitSource {
  id: string
  name: string
  source: string
  owner: string
  provider: string
  repo: string
}

export interface RepokitTemplate {
  description: string
  id: string
  keywords: string[]
  name: string
  path: string
  readme: string
  repoUrl: string
  source: RepokitSource
}

export type RepokitFilter = {
  id: string
  name: string
  keywords: {
    id: string
    name: string
  }[]
}